requirements.txt
streamlit==1.38.0
opencv-python==4.10.0.84
pillow==10.4.0
pandas==2.2.2
numpy==2.1.1
scipy==1.14.1  # Added for advanced mosaic (KDTree)